Revert "Train segmentation model"
