+++ .gitignore
@@ -5,4 +5,3 @@ train/
 validation/
 logs/
 .vscode/
-.tensorboard
