+++ requirements.txt
@@ -6,4 +6,3 @@ numpy==1.25.0
 opencv-contrib-python==4.7.0.72
 fiftyone==0.21.2
 pycocotools==2.0.6
-tensorflow-examples @ git+https://github.com/tensorflow/examples.git@354d31d024463152b3bd1279600bcc818fdd4159
